+++ krampus.py
@@ -1,5 +1,5 @@
 ###############################################################################
-# krampus 0.3.5 - silver iodide
+# krampus 0.3.7
 ###############################################################################
 # [tell a fun krampus tale]
 ###############################################################################
@@ -8,78 +8,82 @@
       # catching exceptions
     # eventually checks for if a resource exists when pulling by id should
       # raise specific exception that invalid jobs not re-queued
+    # Add the option to perform a dry run with no actioning
 ###############################################################################
-import time
 import os
+import time
+
+# Internal imports
 from lib.krampus_tasks import *
 import lib.krampus_logging
 
 
-# kramp it
 class Krampus():
+    # Setup basic things we need and instantiate logger and krampus_tasks
     def __init__(self, region, bucket_name, key, whitelist, krampus_role):
-        # setup some basic things we need
         self.region = region
         self.bucket_name = bucket_name
-        self.key = key # basically the filename
+        self.key = key
         self.whitelist = whitelist
         self.krampus_role = krampus_role
-        # instanitate logger
-        self.klog = KLog(self.bucket_name, "krampus_log_" + str(int(time.time())))
+
+        self.klog = KLog(self.bucket_name, "krampus_log_{0}".format(str(int(time.time()))))
         self.kt = KTask(self.region, self.bucket_name, self.klog, self.whitelist, self.krampus_role)
 
-    # collect our jobs
+    # Collect jobs to populate kt.tasks
     def getTasks(self):
-        # ktask is our friend dot ru
-        self.kt.getTasks(self.key) # should populate kt.tasks
+        self.kt.getTasks(self.key)
 
-    # complete them
+    # Complete tasks, otherwise defer them to be actioned later
     def completeTasks(self):
         for task in self.kt.tasks:
             try:
                 task.complete()
             except Exception as e:
-                # generic except to keep things moving until we have fixed all the bugs in complete methods
-                # add to deferred tasks to try later
+                # TODO: Resolve bugs in resource complete() methods to avoid deferral
                 self.kt.deferred_tasks.append(task.as_json)
-                # alert that there was an issue
-                KLog.log("could not complete task: %s" % str(e), "critical")
+                KLog.log("Unable to complete task: {0}".format(str(e)), "critical")
 
-    # update the tasks
+    # Update tasks, removing completed and keeping deferred
     def updateTaskList(self):
         self.kt.rebuildTaskList()
 
 
+# Collects information required to run, otherwise set to preset values
 def main(event, context):
-    # collect our run information, otherwise just give it something I guess
     if os.getenv("DEFAULT_REGION"):
         region = os.getenv("DEFAULT_REGION")
     else:
         region = "us-east-1"
+
     if os.getenv("KRAMPUS_BUCKET"):
         krampus_bucket = os.getenv("KRAMPUS_BUCKET")
     else:
         krampus_bucket = "krampus-dev"
+
     if os.getenv("TASKS_FILE_KEY"):
         tasks_file_key = os.getenv("TASKS_FILE_KEY")
     else:
         tasks_file_key = "tasks.json"
+
     if os.getenv("ARN_WHITELIST"):
         whitelist = os.getenv("ARN_WHITELIST")
     else:
         whitelist = None
+
     if os.getenv("KRAMPUS_ROLE_NAME"):
         krampus_role = os.getenv("KRAMPUS_ROLE_NAME")
     else:
         krampus_role = "krampus"
+
     # fire it all up
     k = Krampus(region, krampus_bucket, tasks_file_key, whitelist, krampus_role)
     k.getTasks()
     k.completeTasks()
     k.updateTaskList()
-    # save the log file
+
     k.klog.writeLogFile()
-    print "[+] krampus is done sowing death and destruction in AWS... until next time!"
+    print "[+] Krampus is done sowing death and destruction in AWS...until next time!"
 
 
 if __name__ == "__main__":